Add non-SDK version of library Magick++
As a static library Magick++ uses STL and is built with SDK, there is no
way for platform apps or libraries (built without SDK) to refer Magick++
without link-type warning. This adds new static library which is
identical to Magick++, except that it is built without SDK so that
platform apps or libraries can refer it.

Bug: 76228184
Test: mma -j succeeded
Change-Id: I6783bb591c64a7fcc20a722dea759b76127dfe4b